feat(frontend): refactor theme handling, replace with accessibility settings, adjust naming scheme of compoennts, add css for font sizing
