+++ frontend/src/components/TaskCreationDialog.tsx
@@ -13,7 +13,7 @@ import { Label } from '@/components/ui/label';
 import { useState, useEffect } from 'react';
 import { Event } from '@/types/calendar';
 import { addHours, addMinutes, format, setHours, setMinutes } from 'date-fns';
-import { TimePickerForm } from '@/components/time-picker';
+import { TimePickerForm } from '@/components/TimePicker';
 
 interface ConfirmationDialogProps {
 	open: boolean;
